requirements.txt
streamlit==1.51.0
torch==2.5.1
torchvision==0.20.1
pillow>=7.1.0
pandas>=1.4.0
numpy>=1.23
scikit-learn>=1.7.2
matplotlib>=3.10.7
tqdm>=4.67.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
packaging==25.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scipy==1.15.3
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
